Add files via upload
The main purpose of the project was to analyze sales of different businesses over the period of 30 months. For this purpose ETL was performed and data were analyzed using SQL in a python environment and visualized it using matplotlib. Trends in different business, year over year percent different, and a comparison was created. For this purpose time series was used. Most of the businesses showed upward trends in sales. Around 2008 almost all businesses analyzed showed a crash. Book stores sales showed a downward trend in both sales, revenue and sports good showed an upward trend in both cases. Men and women trends showed a downward trend during 2020.